+++ gbx/src/lib.rs
@@ -6,4 +6,5 @@ pub use xml::{base64_encode, Fault};
 mod adapter;
 mod api;
 mod client;
+pub mod file;
 mod xml;