+++ src/database/postgres.rs
@@ -59,7 +59,7 @@ impl PostgresClient {
             UPDATE steward.map
             SET in_playlist = $2
             WHERE uid = $1
-            RETURNING uid, file_name, name, author_login, added_since, in_playlist, exchange_id
+            RETURNING uid, file_name, name, author_login, author_nick_name, added_since, in_playlist, exchange_id
         "#;
         let row = conn.query_opt(stmt, &[&map_uid, &in_playlist]).await?;
         Ok(row.map(Map::from))
@@ -196,7 +196,7 @@ impl Queries for PostgresClient {
     async fn maps(&self) -> Result<Vec<Map>> {
         let conn = self.0.get().await?;
         let stmt = r#"
-            SELECT uid, file_name, name, author_login, author_millis, added_since, in_playlist, exchange_id
+            SELECT uid, file_name, name, author_login, author_nick_name, author_millis, added_since, in_playlist, exchange_id
             FROM steward.map
         "#;
         let rows = conn.query(stmt, &[]).await?;
@@ -207,7 +207,7 @@ impl Queries for PostgresClient {
     async fn playlist(&self) -> Result<Vec<Map>> {
         let conn = self.0.get().await?;
         let stmt = r#"
-            SELECT uid, file_name, name, author_login, author_millis, added_since, in_playlist, exchange_id
+            SELECT uid, file_name, name, author_login, author_nick_name, author_millis, added_since, in_playlist, exchange_id
             FROM steward.map
             WHERE in_playlist
         "#;
@@ -219,7 +219,7 @@ impl Queries for PostgresClient {
     async fn map(&self, map_uid: &str) -> Result<Option<Map>> {
         let conn = self.0.get().await?;
         let stmt = r#"
-            SELECT uid, file_name, name, author_login, author_millis, added_since, in_playlist, exchange_id
+            SELECT uid, file_name, name, author_login, author_nick_name, author_millis, added_since, in_playlist, exchange_id
             FROM steward.map
             WHERE uid = $1
         "#;
@@ -232,12 +232,14 @@ impl Queries for PostgresClient {
         let stmt = r#"
             INSERT INTO steward.map
                 (uid, file_name, file,
-                 name, author_login, author_millis,
-                 added_since, in_playlist, exchange_id)
+                 name, author_login, author_nick_name,
+                 author_millis, added_since, in_playlist,
+                 exchange_id)
             VALUES
                 ($1, $2, $3,
                  $4, $5, $6,
-                 $7, $8, $9)
+                 $7, $8, $9,
+                 $10)
             ON CONFLICT (uid)
             DO UPDATE SET
                 file_name = excluded.file_name,
@@ -252,6 +254,7 @@ impl Queries for PostgresClient {
                     &map.data,
                     &map.metadata.name.formatted.trim(),
                     &map.metadata.author_login,
+                    &map.metadata.author_nick_name.formatted.trim(),
                     &map.metadata.author_millis,
                     &map.metadata.added_since,
                     &map.metadata.in_playlist,
@@ -576,6 +579,7 @@ impl From<Row> for Map {
             file_name: row.get("file_name"),
             name: GameString::from(row.get("name")),
             author_login: row.get("author_login"),
+            author_nick_name: GameString::from(row.get("author_nick_name")),
             author_millis: row.get("author_millis"),
             added_since: row.get("added_since"),
             in_playlist: row.get("in_playlist"),